- Persister updated to support both tenant_cd = and RPAD comparison

+++ src/main/groovy/com/cedarsoftware/ncube/NCubeJdbcPersister.groovy
@@ -91,7 +91,7 @@ class NCubeJdbcPersister
             String app = row.getString('app_cd')
             String version = row.getString('version_no_cd')
             String branch = row.getString('branch_id')
-            ApplicationID appId = new ApplicationID(tenant.trim(), app, version, status, branch)
+            ApplicationID appId = new ApplicationID(tenant, app, version, status, branch)
             cube = buildCube(appId, row)
         })
         if (cube)
@@ -111,7 +111,7 @@ class NCubeJdbcPersister
 /* loadCubeBySha1 */
 SELECT n_cube_id, n_cube_nm, app_cd, version_no_cd, status_cd, revision_number, branch_id, cube_value_bin, test_data_bin, notes_bin, changed, sha1, head_sha1, create_dt
 FROM n_cube
-WHERE ${buildNameCondition(c, 'n_cube_nm')} = :cube AND app_cd = :app AND version_no_cd = :version AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch AND sha1 = :sha1
+WHERE ${buildNameCondition(c, 'n_cube_nm')} = :cube AND app_cd = :app AND version_no_cd = :version AND status_cd = :status AND ${compareTenant()} AND branch_id = :branch AND sha1 = :sha1
 ORDER BY abs(revision_number) DESC""", 0, 1, { ResultSet row ->
                 cube = buildCube(appId, row)
         })
@@ -133,7 +133,7 @@ ORDER BY abs(revision_number) DESC""", 0, 1, { ResultSet row ->
 /* getRevisions */
 SELECT n_cube_id, n_cube_nm, notes_bin, version_no_cd, status_cd, app_cd, create_dt, create_hid, revision_number, branch_id, cube_value_bin, sha1, head_sha1, changed
 FROM n_cube
-WHERE ${buildNameCondition(c, 'n_cube_nm')} = :cube AND app_cd = :app AND version_no_cd = :version AND tenant_cd = RPAD(:tenant, 10, ' ') AND status_cd = :status AND branch_id = :branch
+WHERE ${buildNameCondition(c, 'n_cube_nm')} = :cube AND app_cd = :app AND version_no_cd = :version AND ${compareTenant()} AND status_cd = :status AND branch_id = :branch
 ORDER BY abs(revision_number) DESC
 """, {   ResultSet row -> getCubeInfoRecords(appId, null, records, row) })
 
@@ -282,7 +282,7 @@ VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
             int count = 0
             if (allowDelete)
             {   // Not the most efficient, but this is only used for testing, never from running app.
-                String sqlCmd = "/* deleteCubes */ DELETE FROM n_cube WHERE app_cd = ? AND ${buildNameCondition(c, "n_cube_nm")} = ? AND version_no_cd = ? AND tenant_cd = RPAD(?, 10, ' ') AND branch_id = ?"
+                String sqlCmd = "/* deleteCubes */ DELETE FROM n_cube WHERE app_cd = ? AND ${buildNameCondition(c, "n_cube_nm")} = ? AND version_no_cd = ? AND (tenant_cd = ? OR tenant_cd = RPAD(?, 10, ' ')) AND branch_id = ?"
                 stmt = c.prepareStatement(sqlCmd)
                 for (int i = 0; i < cubeNames.length; i++)
                 {
@@ -800,13 +800,12 @@ VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)""")
                     map.putAll([cube: buildName(c, cubeName), rev: (mustDelete ? maxRev : rollbackRev)])
                     Sql sql = new Sql(c)
 
-                    sql.eachRow(map, """
-                    /* rollbackCubes */
-                    SELECT n_cube_id, n_cube_nm, app_cd, version_no_cd, status_cd, revision_number, branch_id, cube_value_bin, test_data_bin, notes_bin, changed, sha1, head_sha1, create_dt
-                    FROM n_cube
-                    WHERE ${buildNameCondition(c, 'n_cube_nm')} = :cube AND app_cd = :app AND version_no_cd = :version AND status_cd = :status
-                     AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch AND revision_number = :rev
-                     """, 0, 1, { ResultSet row ->
+                    sql.eachRow(map, """\
+/* rollbackCubes */
+SELECT n_cube_id, n_cube_nm, app_cd, version_no_cd, status_cd, revision_number, branch_id, cube_value_bin, test_data_bin, notes_bin, changed, sha1, head_sha1, create_dt
+FROM n_cube
+WHERE ${buildNameCondition(c, 'n_cube_nm')} = :cube AND app_cd = :app AND version_no_cd = :version AND status_cd = :status
+AND ${compareTenant()} AND branch_id = :branch AND revision_number = :rev""", 0, 1, { ResultSet row ->
                         byte[] bytes = row.getBytes(CUBE_VALUE_BIN)
                         byte[] testData = row.getBytes(TEST_DATA_BIN)
                         String sha1 = row.getString('sha1')
@@ -867,7 +866,7 @@ VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)""")
 /* rollbackCubes.findRollbackRev */
 SELECT revision_number FROM n_cube
 WHERE ${buildNameCondition(c, 'n_cube_nm')} = :cube AND app_cd = :app AND version_no_cd = :version AND status_cd = :status
-AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch AND revision_number >= 0 AND sha1 = head_sha1
+AND ${compareTenant()} AND branch_id = :branch AND revision_number >= 0 AND sha1 = head_sha1
 ORDER BY revision_number desc""", 0, 1, { ResultSet row ->
             maxRev = row.getLong("revision_number")
         });
@@ -1066,20 +1065,15 @@ ORDER BY revision_number desc""", 0, 1, { ResultSet row ->
 
         String select = """\
 /* ${methodName} */
-    SELECT n.n_cube_id, n.n_cube_nm, n.app_cd, n.notes_bin, n.version_no_cd, n.status_cd, n.create_dt, n.create_hid, n.revision_number, n.branch_id, n.changed, n.sha1, n.head_sha1
-    ${testCondition}
-    ${cubeCondition}
-    ${notesCondition}
-    FROM n_cube n,
-    ( SELECT n_cube_nm, max(abs(revision_number)) AS max_rev
-    FROM n_cube
-    WHERE app_cd = :app AND version_no_cd = :version AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch
-    ${nameCondition1}
-    GROUP BY n_cube_nm ) m
-    WHERE m.n_cube_nm = n.n_cube_nm AND m.max_rev = abs(n.revision_number) AND n.app_cd = :app AND n.version_no_cd = :version AND n.status_cd = :status AND n.tenant_cd = RPAD(:tenant, 10, ' ') AND n.branch_id = :branch
-    ${revisionCondition}
-    ${changedCondition}
-    ${nameCondition2}"""
+SELECT n.n_cube_id, n.n_cube_nm, n.app_cd, n.notes_bin, n.version_no_cd, n.status_cd, n.create_dt, n.create_hid, n.revision_number, n.branch_id, n.changed, n.sha1, n.head_sha1 ${testCondition} ${cubeCondition} ${notesCondition}
+FROM n_cube n,
+( SELECT n_cube_nm, max(abs(revision_number)) AS max_rev
+ FROM n_cube
+ WHERE app_cd = :app AND version_no_cd = :version AND status_cd = :status AND ${compareTenant()} AND branch_id = :branch
+ ${nameCondition1}
+ GROUP BY n_cube_nm ) m
+WHERE m.n_cube_nm = n.n_cube_nm AND m.max_rev = abs(n.revision_number) AND n.app_cd = :app AND n.version_no_cd = :version AND n.status_cd = :status AND ${compareTenant()} AND n.branch_id = :branch
+${revisionCondition} ${changedCondition} ${nameCondition2}"""
 
         if (max >= 1)
         {   // Use pre-closure to fiddle with batch fetchSize and to monitor row count
@@ -1170,7 +1164,7 @@ ORDER BY revision_number desc""", 0, 1, { ResultSet row ->
     boolean deleteBranch(Connection c, ApplicationID appId)
     {
         Map map = appId as Map
-        new Sql(c).execute(map, "/* deleteBranch */ DELETE FROM n_cube WHERE app_cd = :app AND version_no_cd = :version AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch")
+        new Sql(c).execute(map, "/* deleteBranch */ DELETE FROM n_cube WHERE app_cd = :app AND version_no_cd = :version AND ${compareTenant()} AND branch_id = :branch")
         return true
     }
 
@@ -1195,11 +1189,11 @@ ORDER BY revision_number desc""", 0, 1, { ResultSet row ->
         Map map = appId as Map
         map.putAll([newVer: newSnapVer])
         Sql sql = new Sql(c)
-        sql.executeUpdate(map, "/* releaseCubes */ UPDATE n_cube SET version_no_cd = :newVer WHERE app_cd = :app AND version_no_cd = :version AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id != 'HEAD'")
+        sql.executeUpdate(map, "/* releaseCubes */ UPDATE n_cube SET version_no_cd = :newVer WHERE app_cd = :app AND version_no_cd = :version AND ${compareTenant()} AND branch_id != 'HEAD'")
 
         // Step 2: Release cubes where branch == HEAD (change their status from SNAPSHOT to RELEASE)
         map.create_dt = new Timestamp(System.currentTimeMillis())
-        int releaseCount = sql.executeUpdate(map, "/* releaseCubes */ UPDATE n_cube SET create_dt = :create_dt, status_cd = 'RELEASE' WHERE app_cd = :app AND version_no_cd = :version AND status_cd = 'SNAPSHOT' AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = 'HEAD'")
+        int releaseCount = sql.executeUpdate(map, "/* releaseCubes */ UPDATE n_cube SET create_dt = :create_dt, status_cd = 'RELEASE' WHERE app_cd = :app AND version_no_cd = :version AND status_cd = 'SNAPSHOT' AND ${compareTenant()} AND branch_id = 'HEAD'")
 
         // Step 3: Create new SNAPSHOT cubes from the HEAD RELEASE cubes (next version higher, started for development)
         ApplicationID releaseId = appId.asRelease()
@@ -1267,7 +1261,7 @@ ORDER BY revision_number desc""", 0, 1, { ResultSet row ->
         Map map = appId as Map
         map.putAll([newVer: newVersion, status: 'SNAPSHOT'])
         Sql sql = new Sql(c)
-        int count = sql.executeUpdate(map, "/* changeVersionValue */ UPDATE n_cube SET version_no_cd = :newVer WHERE app_cd = :app AND version_no_cd = :version AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch")
+        int count = sql.executeUpdate(map, "/* changeVersionValue */ UPDATE n_cube SET version_no_cd = :newVer WHERE app_cd = :app AND version_no_cd = :version AND status_cd = :status AND ${compareTenant()} AND branch_id = :branch")
         if (count < 1)
         {
             throw new IllegalArgumentException("No SNAPSHOT n-cubes found with version " + appId.version + ", therefore no versions updated, app: " + appId)
@@ -1293,7 +1287,7 @@ ORDER BY revision_number desc""", 0, 1, { ResultSet row ->
 /* updateTestData */
 UPDATE n_cube SET test_data_bin=:testData
 WHERE app_cd = :app AND ${buildNameCondition(c, 'n_cube_nm')} = :cube AND version_no_cd = :ver
-AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch AND revision_number = :rev"""
+AND status_cd = :status AND ${compareTenant()} AND branch_id = :branch AND revision_number = :rev"""
 
         int rows = sql.executeUpdate(map, update)
         return rows == 1
@@ -1310,7 +1304,7 @@ AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :
         String select = """\
 /* getTestData */
 SELECT test_data_bin FROM n_cube
-WHERE ${buildNameCondition(c, 'n_cube_nm')} = :cube AND app_cd = :app AND version_no_cd = :version AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch
+WHERE ${buildNameCondition(c, 'n_cube_nm')} = :cube AND app_cd = :app AND version_no_cd = :version AND status_cd = :status AND ${compareTenant()} AND branch_id = :branch
 ORDER BY abs(revision_number) DESC"""
 
         sql.eachRow(select, map, 0, 1, { ResultSet row ->
@@ -1342,7 +1336,7 @@ ORDER BY abs(revision_number) DESC"""
 /* updateNotes */
 UPDATE n_cube SET notes_bin = :notes
 WHERE app_cd = :app AND ${buildNameCondition(c, 'n_cube_nm')} = :cube AND version_no_cd = :version
-AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch AND revision_number = :rev""")
+AND status_cd = :status AND ${compareTenant()} AND branch_id = :branch AND revision_number = :rev""")
         return rows == 1
     }
 
@@ -1356,7 +1350,7 @@ AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :
         Sql sql = new Sql(c)
         List<String> apps = []
 
-        sql.eachRow("/* getAppNames */ SELECT DISTINCT app_cd FROM n_cube WHERE tenant_cd = RPAD(:tenant, 10, ' ')", map, { ResultSet row ->
+        sql.eachRow("/* getAppNames */ SELECT DISTINCT app_cd FROM n_cube WHERE ${compareTenant()}", map, { ResultSet row ->
             if (row.getFetchSize() < FETCH_SIZE)
             {
                 row.setFetchSize(FETCH_SIZE)
@@ -1379,7 +1373,7 @@ AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :
         List<String> snapshotVersions = []
         Map<String, List<String>> versions = [:]
 
-        sql.eachRow("/* getVersions */ SELECT DISTINCT version_no_cd, status_cd FROM n_cube WHERE app_cd = :app AND tenant_cd = RPAD(:tenant, 10, ' ')", map, { ResultSet row ->
+        sql.eachRow("/* getVersions */ SELECT DISTINCT version_no_cd, status_cd FROM n_cube WHERE app_cd = :app AND ${compareTenant()}", map, { ResultSet row ->
             if (row.getFetchSize() < FETCH_SIZE)
             {
                 row.setFetchSize(FETCH_SIZE)
@@ -1406,7 +1400,7 @@ AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :
         Sql sql = new Sql(c)
         Set<String> branches = new HashSet<>()
 
-        sql.eachRow("/* getBranches.appVerStat */ SELECT DISTINCT branch_id FROM n_cube WHERE app_cd = :app AND version_no_cd = :version AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ')", map, { ResultSet row ->
+        sql.eachRow("/* getBranches.appVerStat */ SELECT DISTINCT branch_id FROM n_cube WHERE app_cd = :app AND version_no_cd = :version AND status_cd = :status AND ${compareTenant()}", map, { ResultSet row ->
             if (row.getFetchSize() < FETCH_SIZE)
             {
                 row.setFetchSize(FETCH_SIZE)
@@ -1427,7 +1421,7 @@ AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :
     {
         Map map = appId as Map
         Sql sql = new Sql(c)
-        String statement = """/* ${methodName}.doCubesExist */ SELECT DISTINCT n_cube_id FROM n_cube WHERE app_cd = :app AND version_no_cd = :version AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch"""
+        String statement = "/* ${methodName}.doCubesExist */ SELECT DISTINCT n_cube_id FROM n_cube WHERE app_cd = :app AND version_no_cd = :version AND ${compareTenant()} AND branch_id = :branch"
 
         if (!ignoreStatus)
         {
@@ -1449,7 +1443,7 @@ AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :
 
         String select = """\
 /* ${methodName}.maxRev */ SELECT revision_number FROM n_cube
-WHERE ${buildNameCondition(c, "n_cube_nm")} = :cube AND app_cd = :app AND version_no_cd = :version AND status_cd = :status AND tenant_cd = RPAD(:tenant, 10, ' ') AND branch_id = :branch
+WHERE ${buildNameCondition(c, "n_cube_nm")} = :cube AND app_cd = :app AND version_no_cd = :version AND status_cd = :status AND ${compareTenant()} AND branch_id = :branch
 ORDER BY abs(revision_number) DESC"""
 
         sql.eachRow(select, map, 0, 1, { ResultSet row ->
@@ -1564,5 +1558,10 @@ ORDER BY abs(revision_number) DESC"""
             return false
         }
         return Regexes.isOraclePattern.matcher(c.getMetaData().getDriverName()).matches()
+    }
+
+    static String compareTenant()
+    {
+        return " (tenant_cd = :tenant OR tenant_cd = RPAD(:tenant, 10, ' ')) "
     }
 }
